Short & Simple Projects Using C
You are writing in a departmental store given the constraint that you could only sell an item of a variety equivalent to the number of same items sold in previous two days. Calculate the minimum quantity we should sell in each day of 1 week. [The no. of items sold on first two days is 1:1].